+++ resources/views/components/layouts/partials/admin-menu-trigger.blade.php
@@ -1,0 +1,8 @@
+<button
+    type="button"
+    class="avicore-admin-menu-trigger lg:hidden"
+    x-on:click="sidebarOpen = true"
+    aria-label="Abrir menú"
+>
+    <x-ui.icon name="menu" />
+</button>